refactor: split map marker
