+++ tsconfig.json
@@ -16,11 +16,6 @@
     "incremental": true,
     "types": ["kakao.maps.d.ts"]
   },
-  "include": [
-    "next-env.d.ts",
-    "**/*.ts",
-    "**/*.tsx",
-    "component/KakaoMap.tsx"
-  ],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "component/map/KakaoMap.tsx"],
   "exclude": ["node_modules"]
 }
